Decoupling crypto module
