+++ qrl/core/CreateGenesisBlock.py
@@ -2,7 +2,7 @@
 # file LICENSE or http://www.opensource.org/licenses/mit-license.php.
 
 from merkle import sha256
-from block import BlockHeader
+from blockheader import BlockHeader
 
 genesis_info = dict()
 genesis_info['Qf3dadd056aa76b192fcde016521840b401ab031f550552bc9568ad109bc0efeda304'] = 1000000
