+++ tests/conftest.py
@@ -9,4 +9,5 @@
 """
 from __future__ import print_function, absolute_import, division
 
+# noinspection PyUnresolvedReferences,PyPackageRequirements
 import pytest
